requirements.txt
fastapi
redis
sqlalchemy
pydantic
cryptography
uvicorn
pyjwt
